feat(useCleanUpCouponStatusOnUnMount): 체크아웃 페이지가 언마운트시, 쿠폰 선택 및 적용 초기화
